Add axios in vue and vuex
