+++ src/plugins/axios.vuex.js
@@ -1,0 +1,7 @@
+import axios from '../utils/axios'
+
+const axiosVuexPlugin = (store) => {
+  store.$axios = axios
+}
+
+export default axiosVuexPlugin
